* develop: (47 commits)
  DKAnimation serialization data format have been changed. (DKVariant::Data -> DKVariant::StructData)
  iOS keyboard bug fixed!
  DKVertexBuffer::BindStream() bug fixed.
  Update VS2015 project files.
  Half-precision floating-point implemented. (IEEE 754-2008 format) - class DKFoundation::DKFloat16
  no message
  External Library (SQLite3) removed.
  ShrinkToFit() function added. (for array classes)
  Set default method of DKStaticTriangleMeshShape Serializer to DKVariant::StructuredData.
  Update copyright
  no message
  lz4 files added to Android.mk
  - DKVertexBuffer::Decl::offset removed. (It will be calculated on runtime) - DKGeometryBuffer content will be serialized by StructuredData (of DKVariant).
  changed member variable types to unsigned
  DKVariant::ImportXML(), DKVariant::ExportXML(), DKVariant::ImportStream(), DKVariant::ExportStream() now supports DKVariant::VStructuredData type.
  Changed language default types (int, long, long long) to stdint types (int32_t, int64_t, etc.) for portability of various 64bit data models. (ILP64, LP64, LLP64)
  Runtime byte-order test function became inline
  no message
  no message
  DKBuffer::Compress() function supports LZ4, LZ4HC format.
  ...

+++ DK/DKFoundation/DKAtomicNumber64.h
@@ -2,7 +2,7 @@
 //  File: DKAtomicNumber64.h
 //  Author: Hongtae Kim (tiff2766@gmail.com)
 //
-//  Copyright (c) 2004-2014 Hongtae Kim. All rights reserved.
+//  Copyright (c) 2004-2015 Hongtae Kim. All rights reserved.
 //
 
 #pragma once
@@ -20,7 +20,7 @@ namespace DKFoundation
 	class DKGL_API DKAtomicNumber64
 	{
 	public:
-		typedef long long Value;
+		typedef int64_t Value;
 		DKAtomicNumber64(Value initialValue = 0);
 		~DKAtomicNumber64(void);
