+++ DK/DKFoundation/DKThread.h
@@ -2,7 +2,7 @@
 //  File: DKThread.h
 //  Author: Hongtae Kim (tiff2766@gmail.com)
 //
-//  Copyright (c) 2004-2014 Hongtae Kim. All rights reserved.
+//  Copyright (c) 2004-2015 Hongtae Kim. All rights reserved.
 //
 
 #pragma once
@@ -26,7 +26,7 @@ namespace DKFoundation
 	class DKGL_API DKThread
 	{
 	public:
-		typedef unsigned long ThreadId;
+		typedef uintptr_t ThreadId;
 
 		// waiting for join.
 		void WaitTerminate(void) const;
